src/modulemgr: Broke up modulemgr.c into several .c files.
